Rename node.js -> DifferentialNode.js
