+++ DifferentialGrowth/DifferentialNode.js
